backend-python/requirements.txt
pydantic
pydantic-settings
python-dotenv
motor
pymongo
bson